Add barebones Go SDK (#1)
This PR adds a first-pass at a Go SDK for Airplane. It supports outputs along with a few more experimental features -- context cancellation, error handling, and JSON parameter deserialization. We'll keep our version on `0.0.x` for awhile, and so can remove any of these if we decide they are no longer relevant.